Fix ambiguous uint64_t to nlohmann::json conversions
Clang on macOS refuses to choose between multiple matching constructors
(int, size_t, double) when converting a `uint64_t` type implicitly to
a `nlohmann::json` value inside initializer lists, causing build failures.

Applied explicit `static_cast<uint64_t>(...)` to specific `uint64_t`
fields in:
- `src/l2/ailee_sidechain_bridge.h`
- `include/ProverSwarm.h`
- `src/l2/L2State.cpp`

This ensures there is no ambiguity for the compiler, preventing build
failures during the creation of `nlohmann::json` objects.

+++ include/ProverSwarm.h
@@ -62,7 +62,7 @@ struct ProverJob {
             {"job_id", job_id},
             {"payload", payload},
             {"assigned_prover", assigned_prover},
-            {"assigned_at_ms", assigned_at_ms},
+            {"assigned_at_ms", static_cast<uint64_t>(assigned_at_ms)},
             {"completed", completed},
             {"retry_count", static_cast<uint64_t>(retry_count)}
         };
